# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "STExS.Web, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Authenticate/login": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppLoginModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppLoginModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppLoginModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Authenticate/modifyMyProfile": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdentityError"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdentityError"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdentityError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authenticate/register": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppRegisterModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppRegisterModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppRegisterModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Module": {
      "post": {
        "tags": [
          "Module"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleCreateItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleCreateItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleCreateItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Module/{moduleId}": {
      "put": {
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleUpdateItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleUpdateItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleUpdateItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Module/archive": {
      "post": {
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Module/unarchive": {
      "post": {
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Module/getById": {
      "get": {
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailItem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Module/getModulesForUser": {
      "get": {
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDetailItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDetailItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDetailItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Module/getUsersForModule": {
      "get": {
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "name": "moduleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDetailItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDetailItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDetailItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppLoginModel": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AppRegisterModel": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "matrikelNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "IdentityError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ModuleCreateItem": {
        "type": "object",
        "properties": {
          "moduleName": {
            "type": "string",
            "nullable": true
          },
          "moduleDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModuleDetailItem": {
        "type": "object",
        "properties": {
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerFirstName": {
            "type": "string",
            "nullable": true
          },
          "ownerLastName": {
            "type": "string",
            "nullable": true
          },
          "moduleName": {
            "type": "string",
            "nullable": true
          },
          "moduleDescription": {
            "type": "string",
            "nullable": true
          },
          "archivedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isArchived": {
            "type": "boolean",
            "readOnly": true
          },
          "isDeleted": {
            "type": "boolean",
            "readOnly": true
          },
          "chapterIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "isFavorited": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ModuleUpdateItem": {
        "type": "object",
        "properties": {
          "moduleName": {
            "type": "string",
            "nullable": true
          },
          "moduleDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProfileUpdateItem": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "matrikelNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}